apps/TzevaAdom/requirements.txt
requests
paho-mqtt
boto3
botocore